+++ lib/src/controllers/vimeo_player_value.dart
@@ -6,6 +6,7 @@ class VimeoPlayerValue {
   final bool isFullscreen;
   bool isBuffering;
   final bool hasEnded;
+  final bool hasPreloaded;
   final String? videoTitle;
   final double? videoPosition;
   final double? videoDuration;
@@ -19,6 +20,7 @@ class VimeoPlayerValue {
     this.isFullscreen = false,
     this.isBuffering = false,
     this.hasEnded = false,
+    this.hasPreloaded = false,
     this.videoTitle,
     this.videoPosition,
     this.videoDuration,
@@ -32,6 +34,7 @@ class VimeoPlayerValue {
       bool? isFullscreen,
       bool? isBuffering,
       bool? hasEnded,
+      bool? hasPreloaded,
       String? videoTitle,
       double? videoPosition,
       double? videoDuration,
@@ -41,6 +44,7 @@ class VimeoPlayerValue {
     return VimeoPlayerValue(
       isReady: isReady ?? this.isReady,
       isPlaying: isPlaying ?? this.isPlaying,
+      hasPreloaded: hasPreloaded ?? this.hasPreloaded,
       isFullscreen: isFullscreen ?? this.isPlaying,
       isBuffering: isBuffering ?? this.isBuffering,
       hasEnded: hasEnded ?? this.hasEnded,